collector-py/pyproject.toml
[project]
name = "crypto-collector"
version = "2.0.0"
description = "Market data collector with async indicators"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests",
    "ccxt",
    "pandas",
    "psycopg2-binary",
    "python-dotenv",
    "pydantic",
    "pydantic-settings",
    "loguru",
    "prometheus-client",
    "pyyaml",
    "redis",
    "sqlalchemy",
    "selenium",
    "webdriver-manager",
    "cryptography",
    "apscheduler",
    "beautifulsoup4",
    "lxml",
    "curl_cffi",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
pythonpath = ["src"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html:htmlcov",
    "--cov-report=xml",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow tests",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/venv/*",
    "*/env/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
